1.2.9 Mad Libs Assignment
Mad Libs Overview
Mad Libs is a party game invented in 1958. In this game a short story is written with many key words removed and replaced with blanks. Each italicized word provides an indicator for the word that it should be replaced with, for example:

exclamation! she said adverb as she verb (past) her motorized noun and drove off with her adjective cat.
Without hearing the context of the story, a player is asked to come up with words of the correct types. The words are written into the blank spaces and the result is read aloud.

Egad! she said tearfully as she slapped her motorized pool noodle and drove off with her stinky cat.
The story has a completely different meaning every time and the results can often be very silly.

Due: Oct 17, 2025