1.4.4 Coding The Text Adventure Project
Introduction
We've been learning how to use the Text Game Engine by studying the Disco Adventure and Haunted House demo games. If you need a refresher, click here to play any of those demo games!

Project Assignment
In this project we're going to be building our own text adventure game! To do this we're going to be using the game narrative we created in the Project: Narrative Storytelling section, so please make sure you have those documents, especially the flowchart.

We'll then start by downloading and setting up the template files that will act as our base. Click here to download the zip file containing the starter project. If you need help setting up the project refer to The Text Adventure Project Setup section.

Using all of the concepts we've covered in this unit we're then going to build the code for an original text adventure game. If you need help using the game engine refer to the Coding The Text Adventure Game section.

Submit Assignment
Due to the fact that this assignment will contain multiple files it will need to be submitted as a single zip file. To see how to do this for your specific system check the Documentation tab.

Due: Nov 14, 2025