2.2.6 Conditional Story Assignment
Overview
In this assignment we’re going to be creating a short story using code and specifically conditional statements. Using the comparison operators, logical operators, and the conditional statements we’ve learned about in the previous lessons we’ll build a unique story centered around a player escaping a castle. Based on the results of the conditions we test for the story will end up being different.

Due: Dec 23, 2025