2.3.5 Dessert Delegation Assignment
Overview
In this assignment we’re going to be building a dessert delegation system for an imaginary bakery, based on a random ticket that will be issued to our theoretical customer, they will get a specific dessert. We’ll be combining what we’ve learned about conditional statements with what we’ve covered in this lesson regarding the Math object and its methods.