2.5.4 Guess the Word Game Over Logic
Objectives
To get the initial version of the project to the final version there are a few objectives that need to be completed that will fall into three categories.

Setup The Interface
The objective of this section will be to set up the interface, or the screen the player will be interacting with to play the game. The steps to complete this process will be covered in the Guess the Word Game Setup video section.

In Game Logic
The objective of this section will be to code the logic that will allow the player to effectively play the game. The steps to complete this process will be covered in the Guess the Word Game Logic video section.

Game Over Logic
The objective of this section will be to code the logic that will handle if the player wins or loses and restarting the game. The steps to complete this process will be covered in the Guess the Word Game Over Logic video section.

Note, though, that these categories may not extend to every project, and some projects may require additional ways to organize the coding objects.